requirements.txt
asgiref==3.2.10
Django==3.1.1
django-debug-toolbar==2.2
django-environ==0.4.5
flake8==3.8.3
gunicorn==20.0.4
mccabe==0.6.1
Pillow==7.2.0
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
sqlparse==0.3.1
